Create Category and others class update

+++ src/main/java/com/educandoweb/course/repositories/OrderRepository.java
@@ -2,9 +2,7 @@ package com.educandoweb.course.repositories;
 
 
 import com.educandoweb.course.entities.Order;
-import com.educandoweb.course.entities.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-
-public interface OrderRepository extends JpaRepository<Order,Integer> {
+public interface OrderRepository extends JpaRepository<Order, Long> {
 }
